Revert "Move memoryview XID code to Objects/memoryobject.c."
This reverts commit 1239382cc2e48bb91f096de3c822e3c0067d5465.
